fix accelerator related tests
fix types
