+++ frontend/src/__mocks__/mockAcceleratorProfile.ts
@@ -1,34 +1,47 @@
-import _ from 'lodash';
 import { AcceleratorProfileKind } from '~/k8sTypes';
-import { RecursivePartial } from '~/typeHelpers';
-import { TolerationEffect, TolerationOperator } from '~/types';
+import { Toleration, TolerationEffect, TolerationOperator } from '~/types';
+import { genUID } from './mockUtils';
 
-export const mockAcceleratorProfile = (
-  data: RecursivePartial<AcceleratorProfileKind> = {},
-): AcceleratorProfileKind =>
-  _.merge(
+type MockResourceConfigType = {
+  name?: string;
+  namespace?: string;
+  displayName?: string;
+  identifier?: string;
+  description?: string;
+  enabled?: boolean;
+  tolerations?: Toleration[];
+};
+
+export const mockAcceleratorProfile = ({
+  name = 'migrated-gpu',
+  namespace = 'test-project',
+  displayName = 'Nvidia GPU',
+  identifier = 'nvidia.com/gpu',
+  description = '',
+  enabled = true,
+  tolerations = [
     {
-      apiVersion: 'dashboard.opendatahub.io/v1',
-      kind: 'AcceleratorProfile',
-      metadata: {
-        name: 'test-accelerator',
-        annotations: {
-          'opendatahub.io/modified-date': '2023-10-31T21:16:11.721Z',
-        },
-      },
-      spec: {
-        displayName: 'Test Accelerator',
-        enabled: true,
-        identifier: 'nvidia.com/gpu',
-        description: 'Test description',
-        tolerations: [
-          {
-            key: 'nvidia.com/gpu',
-            operator: TolerationOperator.EXISTS,
-            effect: TolerationEffect.NO_SCHEDULE,
-          },
-        ],
-      },
-    } as AcceleratorProfileKind,
-    data,
-  );
+      key: 'nvidia.com/gpu',
+      operator: TolerationOperator.EXISTS,
+      effect: TolerationEffect.NO_SCHEDULE,
+    },
+  ],
+}: MockResourceConfigType): AcceleratorProfileKind => ({
+  apiVersion: 'dashboard.opendatahub.io/v1',
+  kind: 'AcceleratorProfile',
+  metadata: {
+    creationTimestamp: '2023-03-17T16:12:41Z',
+    generation: 1,
+    name,
+    namespace,
+    resourceVersion: '1309350',
+    uid: genUID('service'),
+  },
+  spec: {
+    identifier,
+    displayName,
+    enabled,
+    tolerations,
+    description,
+  },
+});
