+++ frontend/src/api/__tests__/servingRuntimes.spec.ts
@@ -1,10 +1,10 @@
-import { mockAcceleratork8sResource } from '~/__mocks__/mockAcceleratork8sResource';
+import { mockAcceleratorProfile } from '~/__mocks__/mockAcceleratorProfile';
 import { mockServingRuntimeK8sResource } from '~/__mocks__/mockServingRuntimeK8sResource';
 import { mockServingRuntimeModalData } from '~/__mocks__/mockServingRuntimeModalData';
 import { mockServingRuntimeTemplateK8sResource } from '~/__mocks__/mockServingRuntimeTemplateK8sResource';
 import { assembleServingRuntime } from '~/api/k8s/servingRuntimes';
 import { ServingRuntimeKind } from '~/k8sTypes';
-import { AcceleratorState } from '~/utilities/useAcceleratorState';
+import { AcceleratorProfileState } from '~/utilities/useAcceleratorProfileState';
 
 global.structuredClone = (val: unknown) => JSON.parse(JSON.stringify(val));
 
@@ -78,10 +78,10 @@ describe('assembleServingRuntime', () => {
   });
 
   it('should add tolerations and gpu on modelmesh', async () => {
-    const acceleratorState: AcceleratorState = {
-      accelerator: mockAcceleratork8sResource({}),
-      accelerators: [mockAcceleratork8sResource({})],
-      initialAccelerator: mockAcceleratork8sResource({}),
+    const acceleratorProfileState: AcceleratorProfileState = {
+      acceleratorProfile: mockAcceleratorProfile({}),
+      acceleratorProfiles: [mockAcceleratorProfile({})],
+      initialAcceleratorProfile: mockAcceleratorProfile({}),
       count: 1,
       additionalOptions: {},
       useExisting: false,
@@ -96,7 +96,7 @@ describe('assembleServingRuntime', () => {
       mockServingRuntimeK8sResource({ auth: false, route: false }),
       true,
       false,
-      acceleratorState,
+      acceleratorProfileState,
       true,
     );
 
@@ -106,10 +106,10 @@ describe('assembleServingRuntime', () => {
   });
 
   it('should not add tolerations and gpu on kserve', async () => {
-    const acceleratorState: AcceleratorState = {
-      accelerator: mockAcceleratork8sResource({}),
-      accelerators: [mockAcceleratork8sResource({})],
-      initialAccelerator: mockAcceleratork8sResource({}),
+    const acceleratorProfileState: AcceleratorProfileState = {
+      acceleratorProfile: mockAcceleratorProfile({}),
+      acceleratorProfiles: [mockAcceleratorProfile({})],
+      initialAcceleratorProfile: mockAcceleratorProfile({}),
       count: 1,
       additionalOptions: {},
       useExisting: false,
@@ -124,7 +124,7 @@ describe('assembleServingRuntime', () => {
       mockServingRuntimeK8sResource({ auth: false, route: false }),
       true,
       false,
-      acceleratorState,
+      acceleratorProfileState,
       false,
     );
