+++ frontend/src/api/k8s/inferenceServices.ts
@@ -11,7 +11,7 @@ import { InferenceServiceKind, K8sAPIOptions, K8sStatus, KnownLabels } from '~/k
 import { CreatingInferenceServiceObject } from '~/pages/modelServing/screens/types';
 import { translateDisplayNameForK8s } from '~/pages/projects/utils';
 import { applyK8sAPIOptions } from '~/api/apiMergeUtils';
-import { AcceleratorState } from '~/utilities/useAcceleratorState';
+import { AcceleratorProfileState } from '~/utilities/useAcceleratorProfileState';
 import { getModelServingProjects } from './projects';
 import { assemblePodSpecOptions } from './utils';
 
@@ -21,7 +21,7 @@ export const assembleInferenceService = (
   editName?: string,
   isModelMesh?: boolean,
   inferenceService?: InferenceServiceKind,
-  acceleratorState?: AcceleratorState,
+  acceleratorState?: AcceleratorProfileState,
 ): InferenceServiceKind => {
   const { storage, format, servingRuntimeName, project } = data;
   const name = editName || translateDisplayNameForK8s(data.name);
@@ -148,7 +148,7 @@ export const createInferenceService = (
   data: CreatingInferenceServiceObject,
   secretKey?: string,
   isModelMesh?: boolean,
-  acceleratorState?: AcceleratorState,
+  acceleratorState?: AcceleratorProfileState,
 ): Promise<InferenceServiceKind> => {
   const inferenceService = assembleInferenceService(
     data,
@@ -169,7 +169,7 @@ export const updateInferenceService = (
   existingData: InferenceServiceKind,
   secretKey?: string,
   isModelMesh?: boolean,
-  acceleratorState?: AcceleratorState,
+  acceleratorState?: AcceleratorProfileState,
 ): Promise<InferenceServiceKind> => {
   const inferenceService = assembleInferenceService(
     data,
